Add support for sending multicast IP packets.
This commit extends the library to enable the transmission of multicast
IP packets.

Multicast packets differ from unicast packets only in their destination
IP address. The only missing part in the library to support multicast
was the calculation of destination MAC addresses.  Unlike unicast, there's
no need for ARP protocol to find the destination MAC address; it can be
derived directly from the destination IP address.  This commit adds the
missing code for that.